refactor: move is_infrastructure_error to errors.py and cleanup conditionals
