+++ src/copium_loop/constants.py
@@ -37,23 +37,3 @@ MAX_OUTPUT_SIZE = 1024 * 1024
 
 # Default minimum test coverage percentage
 DEFAULT_MIN_COVERAGE = 80
-
-
-def is_infrastructure_error(error_msg: str | None) -> bool:
-    """
-    Identifies common infrastructure/network errors that an LLM cannot resolve.
-    """
-    if not error_msg:
-        return False
-
-    infra_patterns = [
-        "Could not resolve host",
-        "fatal: unable to access",
-        "Connection refused",
-        "Operation timed out",
-        "Network is unreachable",
-        "all models exhausted",
-    ]
-
-    error_msg_lower = error_msg.lower()
-    return any(pattern.lower() in error_msg_lower for pattern in infra_patterns)
