+++ src/copium_loop/nodes/conditionals.py
@@ -8,6 +8,12 @@ from copium_loop.telemetry import get_telemetry
 def should_continue_from_test(state: AgentState) -> str:
     telemetry = get_telemetry()
 
+    if state.get("retry_count", 0) >= constants.MAX_RETRIES:
+        print("Max retries exceeded. Aborting.")
+        telemetry.log_status("tester", "error")
+        telemetry.log_workflow_status("failed")
+        return END
+
     if state.get("node_status") == "infra_error":
         return "tester"
 
@@ -15,18 +21,18 @@ def should_continue_from_test(state: AgentState) -> str:
         telemetry.log_status("tester", "success")
         return "architect"
 
-    if state.get("retry_count", 0) >= constants.MAX_RETRIES:
-        print("Max retries exceeded. Aborting.")
-        telemetry.log_status("tester", "error")
-        telemetry.log_workflow_status("failed")
-        return END
-
     telemetry.log_status("tester", "failed")
     return "coder"
 
 
 def should_continue_from_architect(state: AgentState) -> str:
     telemetry = get_telemetry()
+
+    if state.get("retry_count", 0) >= constants.MAX_RETRIES:
+        print("Max retries exceeded. Aborting.")
+        telemetry.log_status("architect", "error")
+        telemetry.log_workflow_status("failed")
+        return END
 
     if state.get("node_status") == "infra_error":
         return "architect"
@@ -36,12 +42,6 @@ def should_continue_from_architect(state: AgentState) -> str:
     if status == "ok":
         telemetry.log_status("architect", "success")
         return "reviewer"
-
-    if state.get("retry_count", 0) >= constants.MAX_RETRIES:
-        print("Max retries exceeded. Aborting.")
-        telemetry.log_status("architect", "error")
-        telemetry.log_workflow_status("failed")
-        return END
 
     telemetry.log_status("architect", status if status else "error")
     if status == "error":
@@ -53,6 +53,12 @@ def should_continue_from_architect(state: AgentState) -> str:
 def should_continue_from_review(state: AgentState) -> str:
     telemetry = get_telemetry()
 
+    if state.get("retry_count", 0) >= constants.MAX_RETRIES:
+        print("Max retries exceeded. Aborting.")
+        telemetry.log_status("reviewer", "error")
+        telemetry.log_workflow_status("failed")
+        return END
+
     if state.get("node_status") == "infra_error":
         return "reviewer"
 
@@ -61,12 +67,6 @@ def should_continue_from_review(state: AgentState) -> str:
     if status == "approved":
         telemetry.log_status("reviewer", "success")
         return "pr_pre_checker"
-
-    if state.get("retry_count", 0) >= constants.MAX_RETRIES:
-        print("Max retries exceeded. Aborting.")
-        telemetry.log_status("reviewer", "error")
-        telemetry.log_workflow_status("failed")
-        return END
 
     telemetry.log_status("reviewer", status if status else "error")
     if status == "error":
@@ -81,6 +81,12 @@ def should_continue_from_review(state: AgentState) -> str:
 def should_continue_from_pr_creator(state: AgentState) -> str:
     telemetry = get_telemetry()
 
+    if state.get("retry_count", 0) >= constants.MAX_RETRIES:
+        print("Max retries exceeded in PR Creator. Aborting.")
+        telemetry.log_status("pr_creator", "error")
+        telemetry.log_workflow_status("failed")
+        return END
+
     if state.get("node_status") == "infra_error":
         return "pr_creator"
 
@@ -94,12 +100,6 @@ def should_continue_from_pr_creator(state: AgentState) -> str:
         telemetry.log_status("pr_creator", "success")
         return END
 
-    if state.get("retry_count", 0) >= constants.MAX_RETRIES:
-        print("Max retries exceeded in PR Creator. Aborting.")
-        telemetry.log_status("pr_creator", "error")
-        telemetry.log_workflow_status("failed")
-        return END
-
     print(
         f"PR Creator failed or needs commit (status: {repr(status)}). Returning to coder."
     )
@@ -109,6 +109,12 @@ def should_continue_from_pr_creator(state: AgentState) -> str:
 
 def should_continue_from_pr_pre_checker(state: AgentState) -> str:
     telemetry = get_telemetry()
+
+    if state.get("retry_count", 0) >= constants.MAX_RETRIES:
+        print("Max retries exceeded in PR Pre-Checker. Aborting.")
+        telemetry.log_status("pr_pre_checker", "error")
+        telemetry.log_workflow_status("failed")
+        return END
 
     if state.get("node_status") == "infra_error":
         return "pr_pre_checker"
@@ -123,12 +129,6 @@ def should_continue_from_pr_pre_checker(state: AgentState) -> str:
         telemetry.log_status("pr_pre_checker", "success")
         return END
 
-    if state.get("retry_count", 0) >= constants.MAX_RETRIES:
-        print("Max retries exceeded in PR Pre-Checker. Aborting.")
-        telemetry.log_status("pr_pre_checker", "error")
-        telemetry.log_workflow_status("failed")
-        return END
-
     print(
         f"PR Pre-Checker failed or needs commit (status: {repr(status)}). Returning to coder."
     )
@@ -139,9 +139,11 @@ def should_continue_from_pr_pre_checker(state: AgentState) -> str:
 def should_continue_from_journaler(state: AgentState) -> str:
     telemetry = get_telemetry()
 
+    if state.get("retry_count", 0) >= constants.MAX_RETRIES:
+        # Journaler usually doesn't fail in a way that needs retries, but for consistency:
+        return END
+
     if state.get("node_status") == "infra_error":
-        # Journaler is special, it doesn't have a direct loop usually,
-        # but let's retry it anyway if it failed.
         return "journaler"
 
     telemetry.log_status("journaler", "success")
@@ -155,20 +157,21 @@ def should_continue_from_journaler(state: AgentState) -> str:
 def should_continue_from_coder(state: AgentState) -> str:
     telemetry = get_telemetry()
 
-    if state.get("node_status") == "infra_error":
-        return "coder"
-
-    telemetry.log_status("coder", "success")
-    status = state.get("code_status")
-
-    if status == "coded":
-        return "tester"
-
     if state.get("retry_count", 0) >= constants.MAX_RETRIES:
         print("Max retries exceeded from coder. Aborting.")
         telemetry.log_status("coder", "error")
         telemetry.log_workflow_status("failed")
         return END
 
+    if state.get("node_status") == "infra_error":
+        return "coder"
+
+    status = state.get("code_status")
+
+    if status == "coded":
+        telemetry.log_status("coder", "success")
+        return "tester"
+
     print(f"Coder failed (status: {repr(status)}). Retrying...")
+    telemetry.log_status("coder", "failed")
     return "coder"
